+++ agent/anf_agent/protocol.py
@@ -32,6 +32,8 @@ class Verbo(StrEnum):
     VHOST_WRITE = "vhost.write"
     VHOST_REMOVE = "vhost.remove"
     VHOST_LIST = "vhost.list"
+    FSTAB_LIST = "fstab.list"
+    FSTAB_DISABLE = "fstab.disable"
 
 
 class CodiceErrore(StrEnum):